Split custom collections example
Move the Fedify dispatcher setup and bookmark helpers out of main.ts so the
example has clearer module boundaries.  This keeps main.ts focused on the
runnable output while federation.ts owns the Federation configuration and
lib.ts owns the sample data helpers.

https://github.com/fedify-dev/fedify/pull/722#discussion_r3141613216
